move gen-certificate to generator
